// src/app/api/models/dynamic-theme-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DynamicThemeVariantDto } from '../models/dynamic-theme-variant-dto';

/**
 * DTO for dynamic theme information
 */
export interface DynamicThemeDto {

  /**
   * Theme ID
   */
  id?: string;

  /**
   * Whether this theme is built-in
   */
  isBuiltIn?: boolean;

  /**
   * Theme name (e.g., Light, Dark)
   */
  name?: string;

  /**
   * Theme type
   */
  themeType?: string;

  /**
   * Dynamic tokens for this theme
   */
  tokens?: {
[key: string]: string;
};

  /**
   * Available theme variants
   */
  variants?: Array<DynamicThemeVariantDto>;
}